+++ src/DbConnectionPlus/DbConnectionExtensions.QueryOfT.cs
@@ -559,7 +559,7 @@ public static partial class DbConnectionExtensions
     {
         try
         {
-            return (TTarget)ValueConverter.ConvertValueToType<TTarget>(value)!;
+            return ValueConverter.ConvertValueToType<TTarget>(value)!;
         }
         catch (Exception exception) when (value is DBNull)
         {